Add support for filtering internal data format in LiveDataLoader
